+++ application/backend/src/robots/robot_adapter.py
